Also offer to calibrate test results for literal matching strategy.
For some time, the test suite has offered the ability to recalibrate
tags automatically, but no such assistance existed for test using the
literal matching strategy. It is added here.

The literal matching strategy is for test cases that rely on Lintian's
literal output. Common uses cases are for unusual output formats. In
the past, their adjustment was inconvenient.  When the literal output
changed, it had to be copied manually into the test specification.

This commit offers to make these adjustments interactively, similar to
how they were offered for tag-based tests already.