Reduce severity and mark PE32+ related tag experimental; set security features for test binary explicitly.
Two tests for missing security features in PE32+ binaries assumed that
binaries are created with safe standard settings. The settings
recently changed, which caused failures in the the Gitlab pipeline in
unstable.

This commit sets the security features on the binary deemed safe
before disabling them for test purposes. It should restore the Gitlab
CI pipeline for unstable.

Unfortunately, setting the flags does not make the binaries more safe.
The binaries produced by mingw-w64 do not currently implement the
security features they advertise. It actually makes them more
exploitable.

That is why the standard settings were changed. More details can be
found here:

    https://www.kb.cert.org/vuls/id/307144/

Without tools to produce safe binaries, the future of the tag
portable-executable-missing-security-features is in serious doubt.

The tag description recommends setting the flags with genpeimg, but it
is insufficient. The text was amended. The tag's severity was reduced
to pedantic. It was also marked experimental.

Within Debian, the change breaking the CI pipeline was related to the
upload with this changelog entry:

binutils-mingw-w64 (8.8) unstable; urgency=medium

  * binutils 2.34 doesn't fully support enabling the relocation section
    by default, so disable it again, along with the rest of the PE
    protection mechanisms which t make sense without relocation (see
    https://www.kb.cert.org/vuls/id/307144/ for details).

 -- Stephen Kitt <skitt@debian.org>  Wed, 26 Feb 2020 17:50:17 +0100