Add tag and check for absolute symbolic link targets in source trees.
The patched source tree is considered unanchored and can be unpacked
anywhere. It should not contain any absolute paths. Absolute link
targets are also improper. They point outside the source tree.

Sometimes, people point autopkgtest tests toward /dev/null, but that
should be accomplished in another way. Links that are really intended
for installation packages should set the absolute target when building
that package.

This tag was the result of a careful and lengthy rewrite of Lintian's
file index code. It catches a condition that might otherwise lead to
internal inconsistencies.