Transfer collection/md5sums to Index::Md5sums and use it in Index::Patched and Index::Installed.
In the past, md5sums were not a prerequisite for all checks.  The
check deb-format, for example, did not require them. Now md5sums are
always calculated for the installed files and for patched source
trees.

That caused the test deb-format/deb-format-errors-tar to fail with
messages like:

    Command xargs --null --no-run-if-empty md5sum -- exited with status 123:
    md5sum: /usr/share/doc/deb-format-errors-tar/changelog.gz: No such file or directory
    md5sum: /usr/share/doc/deb-format-errors-tar/copyright: No such file or directory

The test is peculiar in that it creates an installation package with
absolute paths. Some investigation was needed as to why the leading
slash remained even though tar produced messages like:

    Removing leading `/' from member names

It was furthermore unclear where these slashes should be removed. In
this commit, are removed before the file names are fed to md5sum. That
means they remain absolute in the index listing. More than likely
absolute paths will have to be filtered in every subsequent
collection.

This commit is part of a series to get rid of the collection databases
and the collection scripts. Ultimately, it will also run checks in
parallel (which they currently do not).

Gbp-Dch: ignore