Provide new test for tag package-contains-timestamped-gzip. (See: #762105)
The previous test modified the mtime of a gzipped archive, but that is
not correct. The member's mtime, which is encoded in the gzip data,
relates solely to the file modification time of the unzipped original.

This test creates gzip files from scratch after setting the file
modification times to the desired values. It also fixes the changelog
time at "Mon, 02 Mar 2020 17:44:19 +0000". Pursuant to Bug#762105,
the changelog timestamp is used as a cutoff for the tag.

Gbp-Dch: ignore