Refocus and rename test offering files ending in .gz that are not actually gzip files.
This test also used to look for package-contains-timestamped-gzip but
the modified only the mtime and not the member's timestamp in the gz
file. Removes the timestamp feature from this test.

Also renames the test for the surviving feature, i.e. files that end
in gz but are not actually gzip files.

The check for compressed files will be rewritten in the near future.
It will come with a specialized test that replaces the functionality
lost here.

Gbp-Dch: ignore