Require collection/changelog-file as a prerequisite for the check files/compressed.
Since the resolution of Bug#762105 via changelog heuristics, the tag
depends on the timing between timestamp and changelog. A rewrite of
collections revealed that the changelog was not actually available
when the check files/compressed ran.

Enables the collection of changelog data for more checks. This change
may have widespread effects for tags that depend on information from
the changelog.