Rewrite check for compressed gzip archives; print modification time of first member. (See: #762105)
This check suffered from several issues. First, it relied on changelog
data but did not require it to be available, That alone calls into
question prior issuance of the tag package-contains-timestamped-gzip.
An earlier commit started to require the changelog data.

Second, the gzip member modification time was extracted incorrectly.
It is in fact a little-endian epoch, and not big-endian. The ->magic
method from File::Path, which is more concise, is now used to read the
first eight bytes.

Third, the check no longer uses file system modification times for any
heuristics.

Fourth, the tag gzip-file-is-not-multi-arch-same-safe, which relates
to architecture, was tied to the member's mtime. That did not make
sense and was decoupled.

The new tag format provides the mtime of the first gzip member as
additional data. It is translated into the UTC time zone and added in
an ISO 8601 format that looks like this:

    2000-02-29Z12:34:56

The format contains no spaces. Perhaps that makes it easy to parse.