Move collections to Processable->unpack for each type; use static dependency handling.
The collections code was much too complex for what it did. This
reduces all of it to about twenty lines of code.

This commit is part of a series to dismantle the great but complex
machinery around collection scripts. For years, Lintian's maintainers
have tried to integrate collections into the regular procedural flow.
Completes the job.