In check debian/copyright, check Excluded-Files only for non-native packages.
After a rewrite of the file indices, the orig index is no longer
available for native packages. One could check the Excluded-Files
field against the patched index for native packages, but somehow that
did not seem to make sense.

Perhaps a tag should be issued if the Excluded-Files field is present
in native packages.