Store file type from file(1) in the index instead of a separate MLDBM database.
Makes the file type as determined by the file(1) magic available for
all regular files in the index. The type can be accessed via the
attribute File::Path->file_info.

A similar method was available before but it relied on a database (and
a code reference, which prevented Processable from being serialized).

The separate database that bundled this information previously is no
longer available.