Store objdump information in the index instead of a gzipped file.
Makes the objdump information previously provided as a gzipped file
available via an attribute in the file index.  It can be accessed with
File::Path->objdump.

Unlike the other collected information, the parser here needs to be
rewritten before all of the the information can be written to the
index. The parser adds entries for libraries (and there is a helper
script in the middle).

The parsing of Processable->objdump_info should be triggered before
attempting to parallelize the checks.

The separate gzip file that bundled this information previously is no
longer available.