Store maintainer script information in the index instead of a separate MLDBM database.
Makes the maintainer script information that was previously available
as a separate database available via an attribute in the file index.
It can be accessed with File::Path->control.

This attribute is currently similar to ->script but not identical. It
would be nice to combined the two in the future.

The separate database that bundled this information previously is no
longer available.