Remove tag pkg-config-references-unknown-shared-library; outside of Lintian's purview. (Closes: #953099)
A pkg-config file can refer to all kinds of libraries. For example, it
was once common to include -lm for the system math library. There is
no way for Lintian to check all build prerequisites.

It is possible someone got confused. The pkg-config ships with other
development files, which do not include the shared object libraries.
There are many things Lintian can check in pkg-config files, but this
is not one of them.

Removes tag from Lintian.