Parse newline escapes correctly in md5sum files. (Closes: #929729)
The previous algorithm assumed that newlines are escaped literally in
md5sum output, but that is not so.

According to the code, newlines are escaped as the two characters \n,
while backslashes are escaped as \\:

/* If ESCAPE is true, then translate each NEWLINE byte to the string, "\\n",
   and each backslash to "\\\\".  */
static void
print_filename (char const *file, bool escape)
{
  if (! escape)
    {
      fputs (file, stdout);
      return;
    }

  while (*file)
    {
      switch (*file)
        {
        case '\n':
          fputs ("\\n", stdout);
          break;

        case '\\':
          fputs ("\\\\", stdout);
          break;

        default:
          putchar (*file);
          break;
        }
      file++;
    }
}

This commit adjusts the parser to read the format correctly.

Sorry about all the noise!