Use Severity field in tags to determine their display prominence directly. (Closes: #935706)
After a discussion in Bug#935706, the old mechanism between tag severity and
tag certainty that determined the warning level was abandoned.

The prominence a tag is given during output is now determined directly by
the field Severity. The value can be one of 'error', 'warning', 'info',
'pedantic' or 'classification'.

The new values correspond directly to the familiar levels, and are often
requested in bug reports.

The old concept of tag certainty no longer exists in Lintian.