Provide test for binary-is-wrong-architecture, which was imported from the xdeb package.
This test worked on amd64 without a hitch when the cross-compiler for
armhf was installed. The build prerequisite was not added to Lintian,
however, since it was not clear if the extra architecture would be
enabled in the Gitlab CI runner.

On amd64 or i386, please follow these steps to run the test:

    dpkg --add-architecture armhf
    apt update
    apt install gcc-arm-linux-gnueabihf

On all other architectures this may work, but it was not tested:

    dpkg --add-architecture amd64
    apt update
    apt install gcc-x86-64-linux-gnu

(Taken from: https://wiki.debian.org/CrossToolchains)

Gbp-Dch: ignore