Analyze Ubuntu's automatic debug packages, which have the non-standard suffix .ddeb. (Closes: #951513, #953857)
Lintian did not previously process automatic debug packages for
Ubuntu. That caused failures in the test suite on Ubuntu after
classification tags were enabled for all tests.

There were several ways to mitigate the issue, but this is probably
the best. Lintian now analyzes Ubuntu's .ddeb packages just like it
does for regular .deb packages.

An initial attempt to solve these bugs showed that Lintian did not
expect a Package-Type field for any installation package. That caused
an error for an invalid field.

A nearby prior commit subsequently made the Package-Type an accepted
field.  Based on Debian policy 5.6.28, a warning is issued if the type
is 'deb', which is the default. Policy discourages that explicit
setting.

This commit (and the preceding ones to deal with the Package-Type
field) were tested on Launchpad. Here is a link to the successful
build log:

    https://launchpadlibrarian.net/469121315/buildlog_ubuntu-eoan-amd64.lintian_2.56.0+git~4_BUILDING.txt.gz

As a small bonus, a merge request was filed with bash-completion
upstream. It should ensure better autocompletion for Ubuntu's .ddeb
file paths when using lintian on them in the future:

    https://github.com/scop/bash-completion/pull/397

Please reopen the bugs if this commit did not work for you.