Substantially refactor Lintian::Profile, Tag::Info; rename CheckScript to Check::Info; drop lintian.desc.
This commit adopts a unified naming scheme for attributes in the three
classes Lintian::Profile, Lintian::Tag::Info and Lintian::Check::Info
(formerly Lintian::CheckScript). For example, checks are now called so
throughout, instead of scripts.

The dummy 'lintian' check for tags issued by Lintian's infrastructure still
exists, but no longer requires a lintian.desc file. The file was removed.

Access to known components instead just enabled ones is now explicit and no
longer hidden in additional arguments to the accessors. The choice between
the original and effective severities is now likewise explicitly stated in
the names of the respective accessors.

The get_tag and get_script accessors are now called get_taginfo and
get_checkinfo. They always return values when the tag or check is known. It
is the user's responsibility to check if the tag (or check) is enabled, if
that information is needed.

Renames Tag::Info->tag to ->name. Eliminates several unnecessary code
references.

This commit should bring no functional changes.

Gbp-Dch: ignore