Fix interaction between two tags about gzip timestamps. (Closes: #954146)
A recent rewrite introduced an error. In packages with multiarch 'same', the
tag for gzip timestamps has a higher severity. It was not a different
condition. This was a genuine bug.

Also fixes test for the affected tag by borrowing from timestamp-in-gzip,
which was similarly broken before commit cea58cc5.

Thanks to Christian Göttsche for bringing it to our attention.