In tag definitions, repurpose the Severity field to indicate an error or warning level directly. (See: #935706)
Reuses the Severity field to state directly the alert level at which the tag
is to be displayed. This abandons an old mechanims which calculated the same
level from the old severity concept and a stated certainty. The certainty no
longer exists.

The Severity field can now have a value of 'error', 'warning', 'info',
'pedantic' or 'classification'. In this commit, the field was set to the
value that was previously issued under the standard Lintian profile.

This commit keeps the Certainty fields in all tag declarations that had
values other than 'certain'. At the time of writing 412 tags were rated
merely 'possible' (out of 1449), while 30 were deemed a 'wild-guess'. Those
tags will be examined at a later date to ensure their diagnostic value meets
Lintian's standards.