In Moo role for checks provide a new, combined entry point 'installable' for deb and udeb.
A new entry point called 'installable' is called for both binary and udeb
packages, but not for any other processable types. It is intended for checks
that currently use the entry point 'always' but whose check type limited
calls to binary and udeb.

As a minor change, calls 'setup' and 'breakdown', which are designed to work
with installable processables in the entry point 'files', only for binary
and udeb.

This commit is part of a series to make check descriptions optional.

Gbp-Dch: ignore