In checks, adjust entry points as preparation for removal of check descriptions.
The check descriptions will soon be optional. They declare a type, however,
that limits when checks are called. Without that limit, the entry point
'always' would sometimes operate on a range of processables that is not
appropriate.

For checks with type 'binary, udeb' the new entry point 'installable' will
make sure only those packages are processed by the check.

This commit is part of a series to make check descriptions optional.

Gbp-Dch: ignore