Record tag changes due to removal of check descriptions; greater scope including buildinfo and changes.
Some checks now run on a wider range of processables. Especially checks for
fields, such as maintainer or uploader, now trigger tags for buildinfo or
changes files.

The test suite contained a udeb without copyright. That condition was not
previously detected.

This commit is part of a series to make check descriptions optional.

Gbp-Dch: ignore