Remove check type from web pages when tags are displayed.
The check type is now determined at run time. Facilities are invoked when
present. Also, checks no longer examine only a single package type. The
concept of a check type is dated.

Removes the check type from tag descriptions generated for web pages. It was
most likely set to an empty string via default in Moo.

This commit is part of a series to make check descriptions optional.

Gbp-Dch: ignore