In output, give overrides a subdued color. (Closes: #954224)
Using a fainter color for overrides as suggested in the bug report produced
a rainbow locally and did not have the intended effect. It made the output
more confusing to the eyes.

This commit implements the second suggestion in the bug report. Overrides
are now shown in 'bright black' when color is selected, although regular
black might have worked just as well.

Thanks to Adam Borowski for bringing this idea to our attention!