Remove temporary directories for pools when they are no longer needed. (Closes: #924449)
From the committer's experience, it is a lot safer to use Path::Tiny's
methods for temporary files and directories than to to mix them with
File::Temp, even though Path::Tiny uses that module internally.

Uses Path::Tiny to create the temporary directory. At least one of the many
attempts to remove it then succeeds. There are no more out of disk errors
locally.

This does not solve the fork bomb problem (Bug#890873) or anything else, but
it stops out-of-disk-space errors locally.