Restore initial working directory before attempting to remove pool.
Various errors during processing can cause Lintian to die. That may leave it
in the pool directory, which then causes this error:

This fix was prompted by e3-2.71, whose orig tarball in the archive is
defective. That caused Index::Orig to abort and subsequently failed to
remove the pool. For details, please see Bug#473092.

Gbp-Dch: ignore