In tests, update expected XML output; remove sed regex for Reproducible Builds.
A previous but nearby commit changed the XML output. This updates the
expected literal output for one test.

Also changes the sed expressions to strip timestamp and Lintian's version.

The architecture for buildinfo files is no longer included (and the
architecture is now an element instead of an attribute). The following
expression was removed:

    //# Under reproducible builds, we also get an architecture dependent
    //# (buildinfo) file.  This causes the architecture field of the changes
    //# file to change.  This little script fixes that.
    s/architecture="source all ([^"]*)"/architecture="source all"/

Please reinsert this snippet if is still needed.