For udebs, disable checks for man pages, copyright and changelog. (Closes: #954428)
The micro packages do not ship man pages, a copyright file, or a changelog.
It makes no sense to look for them, and no warnings should be issued.

The regression occurred when explicit check declarations (*.desc) became
obsolete. They imposed additional restrictions when checks would run. Sorry
about the inconvenience.

Also amended the expected tags for one test. That tag change should have
never been accepted in commit 319fb2ab. It was a mistake of this committer,
who did not realize it was legal to ship code without a coyright notice.

Thanks to Theodore Y. T'so for reporting this issue.