In d/control add lib-cpanel-json-xs-perl as Depends and Build-Depends.
The module JSON::MaybeXS lists the superior Cpanel::JSON::XS as the first
alternative, but pkg-js-tools prevents its automatic installation if it is
installed first.

This ensures that Cpanel::JSON::XS, widely accepted as the best JSON module,
is always available.

For lintian.d.o, a corresponding request was filed as RT#8201.