Simplify mechanism when looking for foreign architectures. (See: #954415)
At first sight, the exceptions to the file(1) patterns seemed related to
binary formats other than ELF on some architectures. That was not the case.

It is most easily seen when running lintian against a foreign .deb package.
That alone produces many false positives. Thanks to Andreas Beckmann for
pointing it out!

This commit drops the exceptions. It should fix the worst offenders for the
hundreds of false positives for this tag.

The xdeb check was not vetted properly being imported from the package with
the same name, although the architecture-related tags are hard to fix
without bug reports. My apologies!

Also fixes the patterns for ARM, although comparing results from file(1) may
not distinguish between the different ARM flavors armel and armhf.

Not yet addressed are the many false positives for multiarch libraries and
cross building toolchains. They require a detailed path analysis (and more
testing) before they can be solved.