Remove xdeb check from Lintian. (Closes: #954415)
The xdeb check was apparently incorporated into the binaries check at
another time. The version there works much better. Removes the check xdeb.

The test written for xdeb is kept. It differs from binaries-from-other-arch
in that it builds a cross binary (even though the build prerequisite was not
added to d/control to avoid burdening Gitlab's CI runs with a relatively
exotic but expensive operation).

This commit rebuilds the profiles but does not incorporate the removal of
the tag gzip-file-is-not-multi-arch-same-safe from the FTP Master auto
reject profile.

That tag is expected to be reinstated soon, pursuant to:

    https://salsa.debian.org/ftp-team/dak/-/commit/3ff66a425d55d888985844f384f09b04c718beea
    https://salsa.debian.org/ftp-team/dak/-/merge_requests/186

and

    https://salsa.debian.org/ftp-team/dak/-/merge_requests/185#note_149893