Check for broken bzip2 files. (See: #702545)
Together with several surrounding commits, this commit checks that
compressed files of seven different algorithms are well-formed.

Unfortunately, the invocation of the respective utilities via xargs, as
suggested in the bug filing, works neither for the utilities nor for the
invocant.

As far as the committer is aware the utilites do not accept more than one
argument. It would also be hard for the check to parse multiple text
exceptions instead of an integer return value.