In check debian/copyright, rationalize global variables and code references; remove auto-correction.
This check used a variety of programming techniques that made it almost
impossible to debug or maintain. For example, the code used Perl's default
search space $_ like a global variable and assigned to it throughout.

This commit also simplifies the use of code references, which can be
extremely hard to debug when they are used in lieu of arrays or other
iterative techniques.

Eliminates auto-correction for typos in d/copyright files. While it is
generally commendable to accomodate minor inaccuracies, doing so is not
advisable in a legal file like a copyright. On the contrary, copyrights
should be examined more strictly than other packaging files.

This prepares the copyright check for a commit series that improves the
handling of UTF-8 files in Lintian.