Look for trailing whitespace with Unicode in mind; issue new tag for text files without newlines at end.
Commit e0f6555c caused a regression in Lintian's own changelog:

    P: lintian source: file-contains-trailing-whitespace debian/changelog (line 17736)
    P: lintian source: file-contains-trailing-whitespace debian/changelog (line 18043)

Fixes string processing when looking for trailing whitespace under
Unicode.

From now on, the whitespace check ignores files that use a national
encoding.  Converts one test case to UTF-8. The additonal tags were
more valuable than the sole tag about the national encoding.

There will be additional regressions from the cruft check. It will be
split into properly working Unicode parts in the near future.

This commit is on the verbose side. It will be condensed when the
check is split.

Gbp-Dch: ignore