Do not strip relative prefixes or extra slashes from wildcards in d/copyright. (Closes: #955538)
The upstream source package for privoxy 3.0.28-stable ships with
awkward filenames. All of them include a dot in the path:

    privoxy-3.0.28-stable/./doc/gpl.html
    privoxy-3.0.28-stable/./doc/pcrs.3
    privoxy-3.0.28-stable/./doc/source/authors.sgml

There are different ways to accommodate this package in Lintian. A
quick discussion follows.

The file indices were rewritten recently with a focus on path
fidelity. It was done to fix several serious bugs arising from unusual
file names. The privoxy package is fascinating proof that the effort
was wildly successful.

The patched index, which is used in all checks that do not explicitly
examine the upstream distribution looks completely normal. There are
no relative prefixes.

The orig index, however, shows the prefixes. They were left in because
paths in the orig tarball are more than just locations. They are
legally relevant pointers for copyright declarations that are not
contained in the files themselves.

In the growing open-source universe, files are often shared and copied
but not always renamed. In fact, much of the time original file are
copied their names are kept intact to indicate where the code came
from. Sometimes, originals are even stored near modified copies to
indicate their provenance.

Because any confusion with respect to such file names could be legally
significant, Lintian's orig index and the code in the copyright check
go through great efforts that no path names are modified. To be sure,
Lintian has internal methods to resolve paths which would remove the
relative prefix at issue here, but it is not wise to use them.

Instead another route was taken. The reporting party asked to please
include the relative prefix in d/copyright. A fully functional copy of
that file was submitted to the bug report for convenience.

On a technical level, this commit stops the removal of relative
prefixes (./) or extra slashes from file names and wildcards listed in
d/copyright. It insures successful matches when the names are listed
exactly the same in both places.

The proposed solution is consistent with other efforts in Lintian and
elsewhere to provide reliable path names for copyright verification.

Due to the extraordinary insights gained from the analysis, a full
path list with all files both shipped and covered by d/copyright in
privoxy 3.0.28-stable, prior to this fix, was attached in the bug
report.