Remove magic reporting placeholders from three tag descriptions.
When printing tag descriptions, Lintian replaces the two placeholders
'&debdev;' and '&maint;' with email addresses for the debian-devel
mailing list and the Lintian maintainers, respectively. The
placeholders were barely used, and are are removed.

The information was not necessary to understand the three tags
affected. It was also common sense in context. Finally, the email
addresses were widely available.

This change simplifies the new reporting framework.

Gbp-Dch: ignore