In File::Path, also provide utf8-based slurping; cache all slurps; use new routines throughout.
With UTF-8 support expanding in Lintian, it became to cumbersome to
examine all files for their UTF-8 validity explicity. This offers a
set of integrated routines.

Slurping is an expensive operation that can lead to excess memory
usage.  For that reason, the results of the new routines are cached.

Gbp-Dch: ignore