Speed up indexing for patched files from source packages.
The substitution style processing for the 'find' output was inspired
by IO::Async::Stream [1], but has serious performance issues.

This commit implements a more straight-forward method of split and
shift to get to the same result. It is much faster.

A reports indicated that lintian took between five and six hours to
run on firefox-esr_68.6.0esr-1~deb9u1.dsc.  The issue is not fully
fixed. A similar issue in md5sums has yet to be addressed.

Thanks to Bastian Blank for indentifying this important issue, and for
contributing substantially to this fix!

[1] https://metacpan.org/pod/IO::Async::Stream#SYNOPSIS