Significantly speed up parsing of output from md5sums.
This a follow-up commit to ea95a064.

The substitution style processing for the 'find' output was inspired
by IO::Async::Stream [1], but has serious performance issues.

This commit implements a more straight-forward method of split and
shift to get to the same result. It is much faster.

A report indicated that lintian took between five and six hours to run
on firefox-esr_68.6.0esr-1~deb9u1.dsc.  The issue is not fully fixed.
A similar issue for file-info data has yet to be addressed.

Thanks to Bastian Blank for identifying this important issue, and for
contributing substantially to this fix!

[1] https://metacpan.org/pod/IO::Async::Stream#SYNOPSIS