Also speed up parsing of magic file-info output from file(1).
This a follow-up commit to ea95a064 and a80d2064.

The substitution style processing for the 'find' output was inspired
by IO::Async::Stream [1], but has serious performance issues.

This commit implements a more straight-forward method of split and
shift to get to the same result. It is much faster.

A report indicated that lintian took up to six hours to run on
firefox-esr_68.6.0esr-1~deb9u1.dsc. After this commit, a similar run
on firefox-esr_68.7.0esr-1.dsc locally takes less than 23 minutes.

More than half of that time seems to be spent executing file(1) via
xargs. The firefox sources ship 299,709 files.

Thanks to Bastian Blank for identifying this important issue, and for
contributing substantially to this fix!

[1] https://metacpan.org/pod/IO::Async::Stream#SYNOPSIS