Drop strip from library; it does not work for Unicode wide character strings and clutters code.
Lintian was recently converted to Unicode. The unpack trick that is
used for speed reasons works only for ASCII characters and not for
Unicode strings:

    https://perldoc.perl.org/functions/pack.html

The routine probably does not work as intended for wide character
strings. It is hereby removed.

Uses of the routine were replaced with a simple substitution
expression throughout. Perl has no standard trim function.

Gbp-Dch: ignore