Downgrade internal uft8 flags on all file system paths delivered by File::Path. (See: #956233, #956723)
This provides relief from runtime errors in Lintian, but does not
solve the bugs. It merely makes Lintian useable again.

The offending packages sphinx and supysonic no longer abort with
runtime errors.

Due to a bug in Perl, strings must be "downgraded" before system calls
such as stat or open. It is the proper fix [1][2], and should happen
in Perl. We simply do so here as triage.

[1] https://github.com/Perl/perl5/issues/10550
[2] https://github.com/Perl/perl5/issues/9674

More comprehensive fixes for both bugs are in the works.