Warn about invalid UTF-8 file names in package sources. (Closes: #956723)
Emits an info-level warning when an orig file tree contains a file
whose name is not valid UTF-8. The name is probably in a national
encoding, but a maintainer may not be able to do much about it.

Repacking may by an option, but it often has other drawbacks, such as
the loss of a cryptographic chain of custody.

Further, issues an error-level message for such a file in the patched
tree when it does not appear in the upstream source. The appearance is
considered the responsibility of the package maintainer.  Here, an
appropriate remedy exists. It is to rename the file.

Includes tests for the new tags.