Exempt lintian from tag patched-file-without-utf8-name; triggered by test case.
The test case for this tag also triggers it in Lintian's own
packaging. Here, we employ our traditional solution. It is to exempt
Lintian programmatically.

There may be other long-term solutions to the situation, but they are
still being explored. This one has the clear advantage that it keeps
Lintian's packaging simple.

Lintian's own packaging is already too magical as it is.

Gbp-Dch: ignore