Fix raw output handle for JSON, which is encoded already; remove duplicate utf8 encoding for regular output.
The combined call to open() that was supposed to duplicate the STDOUT
file descriptor and select a raw encoding at the same time did not
work.  It is a bug in Perl:

    https://github.com/Perl/perl5/issues/7854

But a separate call to binmode worked, as documented here:

    https://stackoverflow.com/a/27802183

The fault was not apparent when Lintian ran in a terminal. It only
became a problem when piping JSON output to a file in the archive tag
sieve. Maybe gnome-terminal reversed the duplicate encoding magically.

Removes a duplicate encoding layer on STDOUT, which was apparently not
a problem here. PerlIO::get_layers reported for $RAW:

    unix perlio encoding(utf-8-strict) utf8 encoding(utf-8-strict) utf8

After this commit it gave the expected values:

    unix perlio