In a test, generate file names that are not valid UTF-8 dynamically.
The unusual file names in the original commit caused problems with
tooling in the FTP archive.

Also removes Lintian's own exemption from the check. No longer
necessary, it was also perceived as hiding the condition from the
archive, where that caused problems per above.

Salsa also choked on the original blob. It displayed a 500 error.