Add new tag for non-UTF-8 file name in native source tree.
The tag patched-file-without-utf8-name was previously issued here, but
that was confusing. The condition was not caused by a patch, after
all, but by a native source file (which like a patched file did not
appear in the orig file list).

For clarity, creates a new tag with the same severity but a different
name.

Includes a test for the new tag. Special care was taken to generate
the non-UTF-8 file names dynamically. Otherwise, they could interfere
with tooling in the FTP archive, and cause a "500 Error" in Salsa.