Fix common prefix detection for extracted tar files. (Closes: #958845)
In this bug, the contents of the simutrans directory were mixed in
with the root directory. In particular, the file

    simutrans/font/m+10r.bdf

was stored in the index as

    font/m+10r.bdf

It only showed up because the mismatch meant that the paragraph in
d/copyright was not being used. Thanks to Jörg Frings-Fürst for
reporting this exciting bug!

A more general description follows.

Many tar files ship paths with a common prefix. It is much safer to
extract a folder rather than multiple files into a current working
directory. The prefix must be removed for Lintian's indexing.

The previous code was flawed. Strikingly, the result depended on the
unsorted order of files examined. The result differed from run to run,
and sometimes was the same. It was non-deterministic.

The new algorithm uses Perl's array processing features to reign in
that random nature. With some luck, it is also correct.