Explain ordering of stand-alone License and Files paragraphs in tag description. (Closes: #959067)
The requirement for placing stand-alone License paragraphs after all
Files paragraphs in d/copyright may not be documented in DEP-5 or
anywhere else, but many examples read that way. To the author's
recollection it has always been like that.

Perhaps it is because there are no other official tools in Debian to
parse d/coyright files that Lintian came to co-define the standard.
The author also remembers having had the problem at least once.

This check is about to be rewritten, but the ordering requirement is
unlikely to be lifted without supporting documentation (or a crowd).
For now, the description is amended, and the bug closed.

Please re-open with more information if you would like to effect a
change in the ordering requirement.