Use aspcud solver in apt like experimental buildds; attempt to restore Gitlab CI.
The new debhelper version 13 from buster-backports is not being
considered when trying to satisfy Lintian's build prerequisites:

    Some packages could not be installed. This may mean that you have
    requested an impossible situation or if you are using the unstable
    distribution that some required packages have not yet been created
    or been moved out of Incoming.
    The following information may help to resolve the situation:
    The following packages have unmet dependencies:
     builddeps:. : Depends: debhelper-compat (= 13)
    E: Unable to correct problems, you have held broken packages.

Thanks to the helpful advice of juliank on #debian-dpkg we considered
pinning to backports, but then settled on his other suggestion to use
the aspcud solver, like experimental buildds do.

This commit cannot be tested in advance. With some luck, it will
restore the Gitlab CI pipeline.