Revert "Make sure apt-get is non-interactive when installing apt-cudf; use cache."
Based on a review of past install logs, I do not think that the
pipeline failures in stable-backports relate to a lack of pinning.

Also, while the aspcud resolver is used in experimental, it may
successfully install packages that are not actually installable. It is
probably not the right approach to restore the pipeline.

This reverts commit 7408c962d95ab58c5b7d05989ffa41c8b99c9bbd.