Revert "Apply change from previous commit to /t/bin/build-test-packages-if-needed, as well."
Based on a review of past install logs, I do not think that the
pipeline failures in stable-backports relate to a lack of pinning.

Also, while the aspcud resolver is used in experimental, it may
successfully install packages that are not actually installable. It is
probably not the right approach to restore the pipeline.

This reverts commit 57e963266fa85439f7668f70d7721e6e97cfe992.