Revert "Use aspcud solver in apt like experimental buildds; attempt to restore Gitlab CI."
Based on a review of past install logs, I do not think that the
pipeline failures in stable-backports relate to a lack of pinning.

Also, while the aspcud resolver is used in experimental, it may
successfully install packages that are not actually installable. It is
probably not the right approach to restore the pipeline.

This reverts commit b2b5de6da567465aa825499283011b4338e2cc42.