Adopt test case from bug for dh sequencer with comments in Makefile recipe (See: #960485)
As a follow-up to commit 595e22d6, provide a test case with the
scenario from the bug report.

This was done as a courtesy to the reporting party to make sure no
future bug filings are required for this matter in the future.

Since the tag is still triggered, commit 595e22d6 did not actually
stop the false positive. That will be fixed in an the next commit.