Add new tag linked-with-obsolete-library; marks only libcblas from the science team for now. (See: !309)
This commit adds a new check and an associated tag for linking with
obsolete libraries. The mechanism is more general than requested.

Currently, the only library detected is libcblas from the Debian
science team.  Please add information and references about offending
libraries to the tag description when extending the check.

Provides a test case for libcblas from the Debian science team.
Regenerates profiles.