Allow for non-Debian profiles to specify the location of Debian VCSes.
It has become more common for Ubuntu packages with permanent deltas to specify
an `XS-Debian-` prefixed field when the Ubuntu package has a separate VCS from
Debian's. This adds those fields to the known source-fields, so
unknown-field-in-dsc is not emitted. A test to confirm this works correctly
has also been added, named debian-vcs-git-in-non-debian-profile.

Further discussion may occur around modifying Debian Policy 5.6.26 to
make this behavior expected and accepted. For now, this is already
implemented downstream, and this simply recognizes it. Further changes
may also occur to prevent reliance on `@if-vendor-is-not`.